MILESTONE: First Test and Framework for Class import/export.
